Various fixes for walk-up registration, including #209/#208, and simplification of redirection logic #206
